refactor web to single page app
