+++ packages/web/tests/App.test.tsx
@@ -37,8 +37,10 @@ vi.mock('@kingdom-builder/engine', () => {
 });
 
 describe('<App />', () => {
-  it('renders Kingdom Builder header', () => {
+  it('renders main menu', () => {
     const html = renderToString(<App />);
     expect(html).toContain('Kingdom Builder');
+    expect(html).toContain('Start New Game');
+    expect(html).toContain('Game Overview');
   });
 });
